Configure the survey data engine

Configures the engine to be used for loading surveys. Checks if the provided engine is supported, sets the default engine if none is specified, and generates a message indicating the configured engine. If the engine is not supported, it throws an error.

Usage

set_engine(.engine = show_engines())

Arguments

.engine

Character vector with the name of the engine to configure. By default, the engine returned by the show_engines() function is used.

Value

Invisibly, the previous engine name (for restoring).

See also

Examples

# \donttest{
old <- set_engine("data.table")
#> Engine: data.table
get_engine()
#> [1] "data.table"
# }